[honeydew][wlan] Remove wlan_core set_region affordance
The `wlan_core` `set_region` affordance was a duplication of the location
set_region affordance. As setting a region is not part of the Fuchsia
WLAN SDK, it doesn't make sense to have an affordance for it.

This change also refactors the corresponding test in `test_wlan_core.py`

Change-Id: I4bbc280937761555c376fe659011ed4e6c23a53f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416936
Reviewed-by: Marc Khouri <mnck@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>