[feedback] Split RebootReason
With the addition of GracefulShutdownAction, it was becoming difficult
to keep track of when the action was/wasn't applicable to RebootReason
and the associated functions in reboot_reason.h/cc. This CL defines an
interface containing most of the functions from the now-deleted
reboot_reason.h.

The interface has two implementations: one for zircon reboot reasons,
and one for graceful reboot reasons. The action is thus only
available/used for the graceful implementation.

Bug: 414413282
Test: `fx test //src/developer/forensics`
Test: `ffx component stop archivist.cm` @ 908973498ebc8df8

[g-review]

Change-Id: I93f8dc4756d850017e616e7b13ec5e632e27c126
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1415521
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Commit-Queue: Thomas Gales <tgales@google.com>