[tracing] Improve performance of TraceFuture
Rust wasn't inlining the TraceFutureExt::trace method. This caused the
future being wrapped to be constructed on the stack and then copied to
the heap. Adding the #[inline(always)] attribute to the method causes
it to be consistently inlined.

Next, the TraceFuture::trace_create call at the end of TraceFuture::new
was preventing TraceFuture::new from being inlined. This caused the
wrapped future to be constructed on the stack and then copied to the
heap as well. I haven't found the create or drop duration events to be
useful when looking at traces and opted to remove them instead of
figuring out how to keep them with inlining. TraceFuture::new is now
consistently inlined when optimizations are enabled. I added the
#[inline(always)] attribute to it as well just to make sure.

With TraceFutureExt::trace and TraceFuture::new inlined, the wrapped
future gets constructed directly on the heap. When the wrapped future
isn't boxed, both the TraceFuture and the wrapped future are constructed
directly inside of the parent future without any copying so I've removed
the Box.

The storage tracing crate now uses the macros from fuchsia_trace when
building for fuchsia and only defines it's own macros when building for
host. This brings in the use of ArgValue::of_registered to storage
tracing.

Trailing comma support has been added to trace_future_args, the fxfs
trace macros, and the storage trace macros.

Fixed: b/456523982
Change-Id: Id198bd936451c8d473e2b4216508059ca5deb02a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421114
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>