[kernel][mp] Disable preemption for local task in mp_sync_exec
There are callers of mp_sync_exec that rely on the calling CPU to not
reschedule while running a task (see https://fxbug.dev/460442928 for
an example).  In the past, a scheduler exemption deferred a local
reschedule when arch_blocking_disallowed was true.  That exemption
is/will-be removed so we need another mechanism to ensure the local
task does not reschedule.  AutoPreemptDisabler is the preferred
mechanism.

mp_sync_exec will now disable preemption when running a task on the
calling CPU and when processing its mailbox.

Add an in-kernel unit test that verifies the new behavior.

Bug: 42143537
Bug: 457759524
Bug: 460442928
Change-Id: I14b5dca052ead1cf62466f63fafdee0779d79fcf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1422178
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>