[storage][f2fs] Replace usage of PAGE_SIZE
Per RFC 0016[1], page size in userspace should be a runtime
determination and this constant should go away.

The f2fs code does not clearly delineate page size from block size, so
as a temporary measure all mention of PAGE_SIZE/kPageSize are replaced
with kBlockSize - and a runtime assert around the relationship is added
to F2fs::Create(), to be revisited if and when we want to make this code
supported in non-4KiB-paged contexts.

[1] https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0016_boot_time_page_sizes

Bug: 42146863
Change-Id: I3dd66aa242c641596b1250de58571573729c011d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421525
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>