[blobfs][fxfs] Add NeedsOverwrite call to BlobCreator.
This is meant to be a way to get the ondisk status of a blob. It could
be This will default to `false` for now to signal that no overwriting is
required until the more complex decision-making can be instrumented in
blobfs.

This protocol is entirely in-tree and can be changed around or removed
if we decide on a different path later.

Bug: 457411364
Change-Id: Iece04fbabb04e562ce6326dcdf56df698e90e568
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1411814
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
API-Review: James Sullivan <jfsulliv@google.com>