[wlan][test] Close tracing Session proxy so tracing socket will close
Stopping a trace will not trigger the socket producing the trace to
close. Dropping the Session proxy is the only way to terminate a trace
to trigger the socket to close.

WLAN's tracing facility was only half-working before this change. By
happenstance, for hw-sim tests, the wlancfg component always panic's
after it exits, and this was causing the Session proxy to close and
write the trace. That of course is not the sort of behavior we should
rely on in our testing.

For FullMAC tests, there was no component to crash and cause the Session
proxy to close. Those tests would hang waiting for tracing socket to
close. This change prevents that hang.

Change-Id: Ie2326fe9745da38ac8fb17eb187aa48f4c8cf2a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1423095
Reviewed-by: John Tan <johntan@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>