[ffx] Resolution only makes sense for products
Some of the Resolution code was written as though it was possible to
"resolve" a query into a target not in product mode. That was mainly
dead code (and/or just led to error-handling later), so this CL
removes such code, clarifying the data model.

On revisiting the code, it turns out that it makes sense to simply
return an Option<ResolutionTarget> when dealing trying to turn
a target into a Resolution -- it is not an error. This CL also cleans
up a bit of that code.

Change-Id: I70c33fa362d329adde67012d30e657fd6a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421522
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>