[netlink] Support SIOCGIFNAME
Support the SIOCGIFNAME which supplies an interface index and requests
an interface name to be written to the provided buffer.

Add ioctl tests for successful and failure cases. Combined the
SIOCGIFINDEX and SIOCGIFNAME calls into a single test, as testing
SIOCGIFNAME standalone would require assuming an interface's id, which
is not stable. This follows the pattern for the equivalent gVisor test.

Fixed: 325639438
Change-Id: Ibf75d897d2b17afea9f3f6635e6d0e83a0426405
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421527
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>