[trace][lacewing] Switch to sync socket
When trace manager receives the stop request, it begins flushing trace
data to a socket, which in turn causes RCS to stall as it fills its
outgoing buffer. In practice, this causes the stop request to stall out,
which results in an eventual ZX_ERR_PEER_CLOSED.

We can mitigate the effects of this while we dig into the internals by
instead starting a background thread that begins reading out of the
socket as soon as the stop request is received, and yields occasionally
to the main thread to handle the stop request.

Bug: 446873535
Change-Id: Idb89c54969f5c2e453f6d35868f6b5ca3cf03466
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421694
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Chris Masone <cmasone@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>