[kernel][vm] Re-validate mapping in PageFault
Changes the locking structure in the page fault path to re-validate
whether the mapping is still valid or not after acquiring the object
lock. This allows for the aspace lock to be dropped prior to acquiring
the VMO lock, instead of after, preventing contention on the VMO lock
from propagating into the aspace lock.

Re-validation of the mapping leverages the property that the VMO in
the mapping is nearly immutable, with the exception of being set to
null when the mapping transitions into the dead state. As this
transition is not reversible and happens under the object lock, it
can be coordinated with an atomic, instead of requiring the aspace
lock.

This change completely removes the aspace lock as a source of lock
contention in page faults for relevant Starnix workloads.

Bug: 452015888
Run-All-Tests: True
Change-Id: Ib1042f9c4361ab05a11d1aa800e5147323e2a35e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420454
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>