[kernel][vm] Process pages before supply
Add a pre-processing function, VmCowPages::ProcessPagesForSupply, to be
called in VmObjectPaged before calling VmCowPages::SupplyPages. This
removes the possibility of waiting on a page request in
VmCowPages::SupplyPages.

Currently the function converts references to pages if there is a user
pager.

Also remove VmPageSpliceList::PeekReference as it it no longer needed.

Bug: 434536251
Run-All-Tests: True
Change-Id: I844e58da1cbc106b1615567caf08fc6716e3c33d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1404174
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Sage Barreda <sagebarreda@google.com>