[kerne][vm] Test for hidden zero page dedupe
Adds unittest for operations on leaf VMOs where a previously allocated
page in a hidden parent has been deduped to the zero page.

Test is intended to avoid regressions in this behavior as it is not
presently tested.

Change-Id: Ib4b8b627f1515af5ea64e54a6e625d512cee9ec4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1404173
Commit-Queue: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Sage Barreda <sagebarreda@google.com>