[kernel][debug][x86] Validate DR6
DR6 is set by the hardware when a #DB occurs to indicate to the
debugger the reason for the exception. As a consequence there should
never be a #DB with no reason in DR6.

Adding this check as a KERNEL_OOPS is part of root causing the linked
bug, where userspace is being told it took a debug exception, but the
DR6 it receives contains no cause.

Bug: 453746441
Change-Id: Iaceac84e9882ed2880689fd1add995f251b9ec3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1425017
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>