[fastboot] Support update-super command
Add support for the `update-super` command and allow it to be used to
determine the partition target when flashing the `blob` volume. By
default, when flashing `blob`, userdata is unaffected. To support
`fastboot flashall`, we can use the `update-super` command to convey to
the fastboot component if we need to wipe userdata or not.

Bug: 458075061
Test: Tested locally on device.
Change-Id: Ifdd8fa7b1fef6ba3faa2400b35fb8082ff6bb132
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1414417
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>