[zircon] Job policy test avoids creating zero-sized physical VMOs
Instead of attempting to create a zero-sized VMO, attempt to create a
1-page VMO.

Even though the creation of this VMO will fail due to the invalid
handle, we are expecting a policy error, not an error related to an
invalid size.

We may soon disallow creating zero-sized physical VMOs. This is related
to our ongoing project of tightening the syscall APIs related to VMO
size.

Bug: 460506492
Change-Id: Ic6ccb4670894f3818eea12a61972be442d2f2cd6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1422517
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Ethan Williams <ethanws@google.com>