[rbe] Support remote fsatrace on C++ compiles
cxx_remote_wrapper.py (and reclient_cxx.sh by extension)
can now work with '--compare --fsatrace-path=...'.

This works by writing the fsatrace command prefix into a
temporary script that is passed to rewrapper --remote_wrapper.
This avoids interference with reclient's built-in C++ input processing.
One limitation is that this is not (yet) composable
with other remote-wrappers, but there are solutions for this.

Fixed: 42079382
Bug: 456902833
Change-Id: Iee9a5bf4759402b9b27737924ef3789e1912a40d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1423654
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Fang <fangism@google.com>