[uart][dw8250] Standalone driver for the DW8250 uart
DW8250 has some extra features and bits that make it somewhat
incompatible (but better) with the plain 8250 driver. In order to keep
the maintenance of the original 8250 a bit simpler, fork a copy of the
driver and customize it and remove dw8250 references from the original.

Change-Id: Ieb7b9a5741aed22be15dfbcdfb6a3944c016ce69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1412275
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Will Drewry <drewry@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>