[fxfs] Write out fsverity tree using f2fs format.
Switching to writing the format of f2fs, mostly to move the digest
information out of the metadata which would be encrypted differently
than the data, possibly leaking information about the file contents.

Since we need to support this format anyways in order to make migration
work with fscrypt, we may as well use that structure here and phase out
the old one.

Bug: 450398331
Change-Id: Ie9e47472f36f0da109bf772f869605dd2ab46c54
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1425414
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>