[netpol] Remove CheckPresence from SocketProxy protocol
This endpoint is unused. We decided to have a Starnix kernel flag and netcfg configuration flag to determine whether these protocols should be used. When these flags are enabled, it is expected that the backing protocol will be present on the product to support those calls. This revert cleans up the unused FIDL endpoint and associated files that handled the calls.

Bug: 385368895
Change-Id: If4a7bb6b0ba392183a365ed965ba5361594ee365
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272004
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>