[internet-checksum] Migrate to criterion
The supported benchmarking framework in fuchsia is `criterion`. The
previous benchmarks in `cargo bench` can never be built in fuchsia
and run against a target. This change also allows the benchmarks to
be run in `perfcompare` builders to monitor the performance changes
in CQ/CI.

The benchmarks themselves are unchanged.

Change-Id: I6a6a696401cf5da192c3116a0ca145e0bd00022b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426034
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>