[wlan] Write beginning of duration in wlan_trace::duration
Instead of only writing the duration event at the end of a duration in
the wlan_trace::duration macro, the macro now always writes a duration
begin event at the beginning of a duration. This prevents ambiguity when
a component unexpectedly exits in the middle of a duration and there is
not duration present in the trace at all.

Change-Id: Icd1f633199c781d4acddee3459360e72f20e5780
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426355
Reviewed-by: John Tan <johntan@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>