[bt] Define a platform configuration for bt-rfcomm
Define a new AIB for bt-rfcomm, which can be optionally added to the
build.

Add a basic platform configuration for RFCOMM. By default, this is
enabled. This matches existing behavior in which it is packaged with
bt-init.

Add the RFCOMM AIB, if it is enabled, in the platform config rules for
Bluetooth.

Add a configuration capability to bt-init which is used to dynamically
route the bredr.Profile API. If enabled, Profile requests will be routed
to bt-rfcomm first. This behavior is unchanged.

Bug: 457743695
Test: fx build, fx ota, verify bt-rfcomm starts
Test: fx test bt-init-unittests
Change-Id: If5308defc5837462a2fc9be0c038c8203a771b46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1412045
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>