[starnix][userfault] ENOTSUP on using userfault without SIGBUS
The only usage of userfault currently supported involves using
UFFD_FEATURE_SIGBUS. Return NOT_SUPPORTED as soon as possible when we
detect that we try to run without it.

Note that we can't error out in UFFDIO_API ioctl directly as running it
with an empty set of features is a supported way to get a list of all
supported features.

Test: UffdTest.*
Bug: 391599171
Change-Id: I7f7a0e567487e9edbf342b20632e58d34067b608
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426295
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Maria Glukhova <mariagl@google.com>