[checksum] Remove `add_bytes_small`
The original implementation had a separate algorithm for calculating the
checksum when the buffer was sufficiently small (< 64 bytes). This was
originally added as an optimization, however current benchmarks reveal
that the main algorithm in `add_bytes` is more efficient, even when the
buffer is small.

Bug: 462085360
Change-Id: I4b679fac3235a30c0074dba8c9005a6cc4427518
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426474
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>