[kernel][x86] Downgrade spurious #DB OOPS to printf
When a spurious #DB exception occurs, x86_debug_handler will emit an
OOPS with some extra details.  This was added to help troublshoot
b/453746441.  These are now happening a lot on CI/CQ so we're
downgrading the printf to an OOPS in order to eliminate this as a
source of CI/CQ failures (think flake rate reduction).

Bug: 453746441
Bug: 461514946
Change-Id: Ic8f494f84c63c9454da0da82c0ca43a02b30f108
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427774
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>