[bazel] bazel_action.py: Remove genquery build and output files
To get rid of a surprising consistency flake, remove the auto-generated
genquery build file and its output file after running the main
`bazel build` command.

This removes a source of flakes on our consistency builders (example at [1]),
that only happens when Ninja decides to schedule two bazel_action()
commands in a different order between two consecutive clean builds.

This flake was very likely the reason for the revert in [2], and it
is hoped here that this will allow re-landing [3] properly.

[1]: https://ci.chromium.org/ui/p/turquoise/builders/global.ci/core.x64-asan-bazel_consistency/b8698204306356068097/overview
[2]: https://fuchsia-review.googlesource.com/c/fuchsia/+/1423316
[3]: https://fuchsia-review.googlesource.com/c/fuchsia/+/1399534

Change-Id: Ifa470254fb8ded5cce3a23eb53158a70c1563982
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427674
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>