[memory_monitor] Detect and break out of cycles
memory_monitor iterates through the graph of VMO parenthood to
propagate memory usage of a VMO down to its leaves, as an accounting
strategy. In the past we have had issues with self-referential VMOs,
who report themselves as their own parents (such as IOBuffer), causing
infinite recursion, which manifest as a stack overflow.

We have already added detection of this specific case as a mitigation,
but we might as well proactively detect anomalous topologies and
make error reporting a bit clearer.

Let's also use the opportunity to fix a bunch of lints in older code,
in no particular order:

  * Use designated initializers when relevant.
  * Fix a bunch of accidental visibility overrides.
  * Swap TRACE_DURATION_BEGIN/END pairs for their scoped equivalent.
  * Rename committed and populated scaled bytes fields (holdover
    cleanups for an earlier change).
  * Minor namespace/header cleanups.

Bug: 377993710, 413664909
Change-Id: Ic9b1a22df0e60d6ef312f66774d6fd99d379bc6b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427794
Commit-Queue: Guillaume Chelfi <chelfi@google.com>
Reviewed-by: Étienne J. Membrives <etiennej@google.com>