[bazel][idk] Add in-tree support to idk_cc_shared_library()
Previously, in-tree targets could not depend on the shared object
generated by idk_cc_shared_library() because it does not have a `CcInfo`
provider. Import the generated shared object so it can be used in-tree.

Also add separate test executables targets for each type of C++ library.

NO_IFTTT=This fixes Bazel behavior and does not need to be replicated in the GN template.

Bug: 417307356
Test: fx bazel build --config=fuchsia //build/bazel/bazel_idk/tests:build_only_tests
Change-Id: I984e9704b01bff26282d36195284f44d38aafbe4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427115
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>