[nmfs] Move filesystem to a module
Common style in Starnix is to place filesystems inside their own module
so that it can be self-contained. Update nmfs to be within its own
module and initialize with the kernel via the "Expando" pattern.

Add test target as well and ensure tests continue to pass. The concept
of a NetworkManagerHandle is now satisfied by the Expando itself so it
is no longer needed.

Bug: 461796926
Change-Id: I0f7a45ad71f239423cc1bc55c72d80633d3e61f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416959
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>