[bisect] Make longest_dimension strategy symmetric.
Previously, the "midpoint" artifact would be removed if the previous
test failed, but it would remain in "remaining_artifacts" if the
previous test passed. This worked, but could result in us re-testing
that artifact down the line (and generally speaking it's weird for it to
behave asymmetrically like this).

Remove midpoint in both test result situations, to prevent accidentally
retesting the same artifact.

Also add product_bundle tests to the root ffx/plugins dependency tree,
and make the search_space tests more relevant.

Change-Id: Ia275c831dec790f8de239053edcd41b6d2088cf5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1422384
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>