[cog] Simplify setup_cog_workspace.py.
This change removes the requirement to set the
`FUCHSIA_ALLOW_SETUP_COG_WORKSPACE` environment variable, instead
replacing it with a check to see whether the script is run within a cog
workspace and redirecting the user to http://go/fuchsia-cog-user-guide
if it is not.

This change also removes some dead code, fixes workspace tests, and
cleans up some of our type hints.

Bug: 459918149
Change-Id: I959aa7767df9081cc8072646df55192a75e7633e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1428254
Commit-Queue: Darren Chan <chandarren@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>