[starnix] Simplify memory attribution
The restricted address space is at a fixed location, with a fixed size.
We thus don't need to dynamically query MemoryManager for this
information when attributing memory to Starnix processes.

Not requiring MemoryManager allows us to simplify the notification code,
as we can start sending the attribution as soon as the pid is known,
instead of waiting for MemoryManager to be instantiated.

Change-Id: Ie7cd7f7468340c88fe4129c68468c4ba5cdbdde6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416554
Commit-Queue: Étienne J. Membrives <etiennej@google.com>
Reviewed-by: Wez <wez@google.com>