[trace][Lacewing] Catch and ignore FIDL errors when stopping tracing
In the past, investigations into getting ZX_PEER_CLOSED while attempting
to stop tracing indicated that FuchsiaController was no longer able to
communicate with the device. Current evidence suggests that is no longer
the case, and so it's worth trying to catch and warn on this failure
instead of allowing it to escape and abort the test.

In all cases that we've examined, the trace_manager on the host side
receives the Stop FIDL message, flushes the trace back to the host, and
then responds. FIDL indicates to the trace_manager that the response was
sent successfully.

As long as Lacewing is actually able to read the trace, and is able to
send future FIDL messages without incident, we don't actually care too
much about this failure.

Bug: 446873535
Test: run Lacewing tests that use tracing via fx test-remote
Change-Id: I1e290c6a1162ca39681e249566f7ee6229723cc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427994
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>