[fxtest] Allow Gemini API key to be passed to analysis script
This change updates the fxtest tool to correctly pass the GEMINI_API_KEY
environment variable to the Gemini analysis script when the
--gemini-analysis flag is used. It involves propagating the key through
the hermetic environment, storing it in the execution environment, and
passing it as a command-line argument to the analysis script. The help
text for --gemini-analysis was also updated to reflect the correct way
to set the API key, making it so users no longer have to pass it in
using --env.

Test: Updated unit tests to reflect change

Change-Id: I7bc822aeea9e8702b0137d3dab0a43f90aebd1c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427028
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Hardik Singh <singhhardik@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>