[nmfs] Remove reconnection logic and mutex
The NetworkManager had logic to reconnect to the StarnixNetworks
protocol if the state between the socketproxy and the NetworkManager
ever got out of sync or the socketproxy Fuchsia component
crashed/restarted. This behavior has not been witnessed in builds and is
therefore likely not needed, as removing it simplifies the
NetworkManager data structure greatly.

Simplify the constructor by passing in the proxy and an inspect node for
builds where the network_manager is enabled and allow for the default
constructor when the network_manager is initialized from the filesystem.
In the next CL, the functionality will be updated to only instantiate
the NetworkManager when the network_manager flag is enabled rather than
instantiating it from the filesystem calls.

Bug: 461796926

Change-Id: I02eee52d8d02f7cfde2318e954b9083a4ebd2c25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1423436
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>