[fxtest] Disable timeouts when debugging
This overrides any timeout set on the command line, and the default case
(where it is not specified) by setting the timeout value to zero.
Setting the timeout value ourselves also overrides tests that set it
explicitly in their test_specs.

This is the naive approach, but in https://fxbug.dev/462419969 we should
think of a better way to handle this so that the timeout case (when
reproducible at a developer's desk) is catchable in the debugger like
other failures.

Bug: 462419969
Change-Id: Ic817a10718fe0715d23ad292856d5feb6a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1428934
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>