[usb_vsock_host] Clean up randomly-allocated ports
When we initiate a connection from the VSOCK host, that connection
needs a source port and a destination port. The destination port will
be provided by the caller since it identifies what service we're
connecting to, but the source port is randomly generated and usually of
no interest to the user.

When we pick that random port, we mark it as reserved. When the
connection dies, we *should* mark it as free again. This is a bit hard
to do in our code and not a huge problem given the size of the port
space we have set aside for connections, but it's better to do for
cleanliness's sake.

Fixed: 407622199
Change-Id: I6a6a6964851b49246b76086a3502eeb65af23d7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1428135
Commit-Queue: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>