[test][revive] Add an "opt-out" annotation
For tests that have test-case-revive configured, we want to switch
from requiring an "opt-in" annotation on test cases that should be
revived, to having that the default and using opt-out on cases that
should not be revived.

This needs a soft transition, so for now, add the `opt_out` decorator
but still require the `tag_test` decorator.

Also, nerf the `_tag_name` parameter - simply using `tag_test` is
enough of an opt-in.

Test apps that aren't built with `params.test_case_revive = true`
will continue to run normally without revive, regardless of annotation.

Bug: 460154311
Change-Id: Ic7e42da3d9a274d90e6af58f1a587041bd6eeae0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420880
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Commit-Queue: Chris Phoenix <cphoenix@google.com>