[nmfs] Only instantiate NetworkManager with flag
By using "peek" in the Expando, we will only initialize the
NetworkManager struct when the "network_manager" container flag is
enabled in features.rs. Return EPERM on each of the filesystem calls
when the NetworkManager is not present.

The existing tests are split into the proxy enabled and non-proxy
enabled variant. Integrate any checks that are done internal to the
NetworkManager into the proxy variant of the tests and remove the
non-proxy variants.

Bug: 461796926
Change-Id: I27bef5f92826873267bd5b11b2efa85c22ba4279
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426775
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>