[kernel] Print thread pointers in arch_dump_exception_context()
This makes the arch_dump_exception_context() for arm64 and x86
print the thread pointer registers that are not stored in the
iframe.  These are directly in the CPU state for the current
thread (only saved in arch_thread at context switch, not kernel
entry), so make it clear that arch_dump_exception_context() is
always called for an iframe from the current thread.

Change-Id: Id8556d90df38d55641401a7e8c26cbaa502dd8ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1429894
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>