[debugger] Fix incorrect Filter->Attach Configurations
Before there were filters with a setting called "never_attach", which
populated the "weak" field of the corresponding Attach configuration.
This was incorrect, because it resulted in actually claiming exception
channels of things below a job when using recursive, job-only filters,
which is highly undesirable when there are many processes and threads
being spawned within that job.

This mistake caused DebugAgent to be doing _significantly_ more work
than it should have been doing, especially in the cases where multiple
test suites are running in parallel and multiple DebugAgent instances
are involved. This caused specific DebugAgent instances to not receive
certain process or thread creation notifications that are expected to be
received before the process or thread is terminated, which tripped an
assert.

Now the AttachConfig has a new "priority", which properly expresses the
differences between the three Attach configurations we have today:

  1. "no exception channel" (kMinimal priority)
  2. "exception channel but do not send modules" (kWeak priority)
  3. "claim exception channel and send modules" (kStrong priority)

This makes it trivial for us to understand how to raise and lower
priorities based on filter configuration settings to not make this
mistake.

Fixed: 460490217
Change-Id: I0441a7d9b612c7b10319cafa85f63d9b6a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1429134
Reviewed-by: Darren Chan <chandarren@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>