[ffx][repository] Return user error in deregister if repo name prefix
This fixes the situation that ffx reports an internal tool bug when the
repo name for deregister is specified on the command line like 'ffx
target repository deregister fuchsia-pkg://repo-name'. This may happen
in manual use cases or when a script parses the output of pkgctl without
removing the prefix. Instead a user error is returned, explaining the
problem to the user.

Fixed: 462716616
Change-Id: I03cccaa607a2fd433eff1599728339eee612ef37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1430695
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>