[ffx][trace] Fix symbolization crash
This corrects a misunderstanding I had about the
SessionParser::parsed_bytes field.

I thought it held the raw
bytes for the TraceRecord that was most recently returned by
SessionParser::next(). In fact it holds the raw bytes for all of the
RawTraceRecords between the previously returned TraceRecord and the
current one (including the bytes for the new one. This is because not
every RawTraceRecord maps to a TraceRecord. Some RawTraceRecords like String and Thread
interning records are handled by the SessionParser, so that trace
consumers don't have to worry about these implementation details.

The code has been updated so that in addition to setting the
parsed_bytes, the parse_next() function also sets the record_range field
to the range within parsed_bytes for the actual record that was
returned. This allows the ffx trace process_event_records function to
pass just the bytes related to the TraceRecord into the processing
function and preserve the other bytes.

Fixes: 462217132
Change-Id: I1da00818c254790427c41028e242bc33ce9192ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1432774
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Chris Masone <cmasone@google.com>